Wha, some gcc versions are much more pickier than others in their int promotion rules
git-svn-id: https://urjtag.svn.sourceforge.net/svnroot/urjtag/trunk@1604 b68d4a1b-bc3d-0410-92ed-d4ac073336b7

+++ urjtag/src/bus/ejtag.c
@@ -344,7 +344,7 @@ ejtag_bus_init (urj_bus_t *bus)
         printf ("EJTAG version: 3.1\n");
         break;
     default:
-        printf ("EJTAG version: unknown (%d)\n", EJTAG_VER);
+        printf ("EJTAG version: unknown (%lu)\n", (long unsigned) EJTAG_VER);
     }
     printf ("EJTAG Implementation flags:%s%s%s%s%s%s%s\n",
             (BP->impcode & (1 << 28)) ? " R3k" : " R4k",
